update action to find pr number - vs directly commenting
